#336 cell_abs:Add new parameter `multiple` and set default as FALSE so it won't affect other functions using it. When `multiple` is TRUE, it calls a function `para_wp`.
para_wp: Basically, `para_wp` and `para_wr` are from original function `para`. In this function, it calls `para_wr` to insert "<w:r>" tags, which allow multiple inputs. The style attached in "<w:p>" will be assigned in this function, such as alignment and indentation.

para_wr: Styles in "<w:rPr>" will be assigned in this function, such as font size, bold, and italics.